--- frontend/src/Components/CreateCardForm/RatingSelector/RatingSelector.scss
@@ -1,8 +1,0 @@
-.accent {
-  fill:  #F5C518;
-  stroke: transparent !important;
-}
-
-svg {
-  cursor: pointer;
-}
