+++ frontend/src/Components/RatingCustom/RatingCustom.js
@@ -3,7 +3,7 @@ import * as React from "react";
 import { Typography, Rating, Box } from "@mui/material";
 import StarIcon from "@mui/icons-material/Star";
 
-import { useFormCreateCard, useFormCreateCardUpdate } from "../../../Context/FormCreateCardContext";
+import { useFormCreateCard, useFormCreateCardUpdate } from "../../Context/FormCreateCardContext";
 
 const labels = {
   0: "Insuportável (0)",
